refactor(suite): migrate earn modals to account-first flow and consolidate shared utils
