+++ packages/suite/src/components/earn/EarnDashboard/common/EarnDashboardSection.tsx
@@ -4,10 +4,11 @@ import { Translation, TranslationKey } from '@suite/intl';
 import { Badge, BadgeIntent } from '@trezor/components';
 
 import { DashboardSection } from 'src/components/dashboard';
-import { PoweredByBadge } from 'src/components/earn';
-import { type EarnProviderId } from 'src/components/earn/providers/providerMetadata';
 import { useSelector } from 'src/hooks/suite';
 import { selectRouteName } from 'src/reducers/suite/routerReducer';
+
+import { PoweredByBadge } from '../../providers/PoweredByBadge';
+import { EarnProviderId } from '../../providers/providerMetadata';
 
 type EarnDashboardSectionProps = {
     titleId: TranslationKey;
