+++ packages/suite/src/components/earn/index.ts
@@ -2,10 +2,13 @@ export { EarnInANutshellModal } from './modals/EarnInANutshell/EarnInANutshellMo
 export { StakingEarnInANutshellModal } from './modals/EarnInANutshell/StakingEarnInANutshellModal';
 export { YieldEarnInANutshellModal } from './modals/EarnInANutshell/YieldEarnInANutshellModal';
 export { UpdateEarnInANutshellModal } from './modals/EarnInANutshell/UpdateEarnInANutshellModal';
+export { EarnClaimModal } from './modals/EarnClaimModal/EarnClaimModal';
 export { EarnProviderConsentModal } from './modals/EarnProviderConsent/EarnProviderConsentModal';
 export { StakingEarnProviderConsentModal } from './modals/EarnProviderConsent/StakingEarnProviderConsentModal';
 export { YieldEarnProviderConsentModal } from './modals/EarnProviderConsent/YieldEarnProviderConsentModal';
 export { UpdateEarnProviderConsentModal } from './modals/EarnProviderConsent/UpdateEarnProviderConsentModal';
+export { EarnSupplyModal } from './modals/EarnSupplyModal/EarnSupplyModal';
+export { EarnWithdrawalModal } from './modals/EarnWithdrawalModal/EarnWithdrawalModal';
 export { EarnDashboard } from './EarnDashboard/EarnDashboard';
 
 export { EarnSupplyingInfo } from './modals/EarnInANutshell/components/EarnSupplyingInfo';
@@ -15,3 +18,7 @@ export { VotingDelegations } from './VotingDelegations/VotingDelegations';
 export { VotingDelegationsOptions } from './VotingDelegations/VotingDelegationsOptions';
 
 export { PoweredByBadge } from './providers/PoweredByBadge';
+
+export { getStakingHelpCenterLink } from './utils/getStakingHelpCenterLink';
+export { getEarnProviderName } from './utils/getEarnProviderName';
+export { getApyPercent, getApyRate } from './utils/earnApyUtils';
