+++ packages/suite/src/components/earn/modals/EarnInANutshell/EarnInANutshellModal.tsx
@@ -1,49 +1,55 @@
 import { useEffect } from 'react';
 
-import { EarnAccountRef, EarnFlow, EarnProvider } from '@suite-common/suite-types/src/staking';
+import { EarnAnalyticsStep, EarnFlow, EarnProvider } from '@suite-common/suite-types/src/staking';
+import { Account } from '@suite-common/wallet-types';
 import { exhaustive } from '@trezor/type-utils';
+
+import { earnFlowToEventTypeMap } from 'src/constants/suite/staking';
+import { useAnalytics } from 'src/support/useAnalytics';
 
 import { StakingEarnInANutshellModal } from './StakingEarnInANutshellModal';
 import { UpdateEarnInANutshellModal } from './UpdateEarnInANutshellModal';
 import { YieldEarnInANutshellModal } from './YieldEarnInANutshellModal';
-import { useEarnModalAccount } from '../common/useEarnModalAccount';
 
 interface EarnInANutshellModalProps {
     flow: EarnFlow;
     provider: EarnProvider;
-    onCancel: () => void;
-    account?: EarnAccountRef;
+    account: Account;
+    analyticsStep: EarnAnalyticsStep;
     yieldId?: string;
     tokenContractAddress?: string;
+    onCancel: () => void;
 }
 
 export const EarnInANutshellModal = ({
     flow,
     provider,
-    onCancel,
     account,
+    analyticsStep,
     yieldId,
     tokenContractAddress,
+    onCancel,
 }: EarnInANutshellModalProps) => {
-    const selectedAccount = useEarnModalAccount({ account, shouldSyncSelectedAccount: true });
+    const analytics = useAnalytics();
 
     useEffect(() => {
-        if (!selectedAccount) {
-            onCancel();
-        }
-    }, [selectedAccount, onCancel]);
-
-    if (!selectedAccount) {
-        return null;
-    }
+        analytics.report({
+            type: earnFlowToEventTypeMap[flow],
+            payload: {
+                action: 'continue',
+                step: analyticsStep,
+                networkSymbol: account.symbol,
+            },
+        });
+    }, [account.symbol, analytics, analyticsStep, flow]);
 
     switch (flow) {
         case EarnFlow.Stake:
             return (
                 <StakingEarnInANutshellModal
+                    account={account}
                     onCancel={onCancel}
                     provider={provider}
-                    accountRef={account}
                     yieldId={yieldId}
                     tokenContractAddress={tokenContractAddress}
                 />
@@ -51,9 +57,9 @@ export const EarnInANutshellModal = ({
         case EarnFlow.Yield:
             return (
                 <YieldEarnInANutshellModal
+                    account={account}
                     onCancel={onCancel}
                     provider={provider}
-                    accountRef={account}
                     yieldId={yieldId}
                     tokenContractAddress={tokenContractAddress}
                 />
@@ -61,9 +67,9 @@ export const EarnInANutshellModal = ({
         case EarnFlow.UpdateProvider:
             return (
                 <UpdateEarnInANutshellModal
+                    account={account}
                     onCancel={onCancel}
                     provider={provider}
-                    accountRef={account}
                     yieldId={yieldId}
                     tokenContractAddress={tokenContractAddress}
                 />
