+++ packages/suite/src/components/earn/modals/EarnInANutshell/components/EarnSupplyingInfo.tsx
@@ -10,15 +10,16 @@ import {
     SOLANA_EPOCH_DAYS,
 } from '@suite-common/wallet-constants';
 import { selectPoolStatsApyData, selectValidatorsQueueData } from '@suite-common/wallet-core';
+import { Account } from '@suite-common/wallet-types';
 import { BulletList } from '@trezor/components';
 
 import { useSelector } from 'src/hooks/suite';
-import { CoinjoinRootState } from 'src/reducers/wallet/coinjoinReducer';
 import { formatApyValue } from 'src/views/wallet/staking/utils/formatStakeValues';
 
 import { EarnInfoRow } from './EarnInfoRow';
 
 interface EarnSupplyingInfoProps {
+    account: Account;
     isExpanded?: boolean;
     flow: EarnFlow;
 }
@@ -177,14 +178,10 @@ const CardanoSupplyingRows = ({ flow, isExpanded, apy }: EarnSupplyingRowsProps)
     </>
 );
 
-export const EarnSupplyingInfo = ({ isExpanded, flow }: EarnSupplyingInfoProps) => {
-    const { account } = useSelector((state: CoinjoinRootState) => state.wallet.selectedAccount);
-
+export const EarnSupplyingInfo = ({ account, isExpanded, flow }: EarnSupplyingInfoProps) => {
     const validatorsQueue = useSelector(state => selectValidatorsQueueData(state, account?.symbol));
 
     const apy = useSelector(state => selectPoolStatsApyData(state, account));
-
-    if (!account) return null;
 
     const daysToAddToPoolInitial = getDaysToAddToPoolInitial(validatorsQueue);
     const displaySymbol = getNetworkDisplaySymbol(account.symbol);
