+++ packages/suite/src/components/earn/modals/EarnProviderConsent/hooks/useEarnProviderConsentActions.ts
@@ -1,21 +1,23 @@
-import { EarnAccountRef, EarnFlow } from '@suite-common/suite-types/src/staking';
+import { EarnFlow } from '@suite-common/suite-types/src/staking';
+import { NetworkSymbol } from '@suite-common/wallet-config';
 import {
     DEFAULT_VOTING_OPTION,
     selectVotingDelegationOption,
     stakeActions,
 } from '@suite-common/wallet-core';
+import { Account } from '@suite-common/wallet-types';
 
 import { openModal } from 'src/actions/suite/modalActions';
 import { earnFlowToEventTypeMap } from 'src/constants/suite/staking';
 import { useDispatch, useSelector } from 'src/hooks/suite';
-import { selectSelectedAccount } from 'src/reducers/wallet/selectedAccountReducer';
 import { useAnalytics } from 'src/support/useAnalytics';
 
 interface UseEarnProviderConsentActionsProps {
     flow: EarnFlow;
     onCancel: () => void;
     includeVotingDelegation?: boolean;
-    accountRef?: EarnAccountRef;
+    account: Account;
+    networkSymbol?: NetworkSymbol;
     yieldId?: string;
     tokenContractAddress?: string;
 }
@@ -24,13 +26,13 @@ export const useEarnProviderConsentActions = ({
     flow,
     onCancel,
     includeVotingDelegation = false,
-    accountRef,
+    account,
+    networkSymbol,
     yieldId,
     tokenContractAddress,
 }: UseEarnProviderConsentActionsProps) => {
     const dispatch = useDispatch();
     const analytics = useAnalytics();
-    const selectedAccount = useSelector(selectSelectedAccount);
     const selectedVotingDelegation = useSelector(selectVotingDelegationOption);
 
     const report = (action: 'continue' | 'cancel') => {
@@ -39,7 +41,7 @@ export const useEarnProviderConsentActions = ({
             payload: {
                 action,
                 step: 'funds-maintained-modal',
-                networkSymbol: selectedAccount?.symbol,
+                networkSymbol,
                 ...(includeVotingDelegation
                     ? { votingDelegation: selectedVotingDelegation.type }
                     : {}),
@@ -53,7 +55,7 @@ export const useEarnProviderConsentActions = ({
             openModal({
                 type: 'supply',
                 flow,
-                account: accountRef,
+                account,
                 yieldId,
                 tokenContractAddress,
             }),
