+++ packages/suite/src/components/earn/modals/EarnWithdrawalModal/EarnWithdrawalModal.tsx
@@ -1,47 +1,82 @@
-import { useEffect } from 'react';
+import { events } from '@suite/analytics';
+import { Translation } from '@suite/intl';
+import { EarnFlow } from '@suite-common/suite-types/src/staking';
+import { getNetworkDisplaySymbol } from '@suite-common/wallet-config';
+import { Account } from '@suite-common/wallet-types';
+import { CollapsibleBox, Column, Grid, H3, Modal } from '@trezor/components';
 
-import { EarnAccountRef } from '@suite-common/suite-types/src/staking';
-import { getNetwork } from '@suite-common/wallet-config';
-import { Account, SelectedAccountLoaded } from '@suite-common/wallet-types';
+import { WithdrawalFormContext, useWithdrawalForm } from 'src/hooks/earn/useWithdrawalForm';
+import { useLayoutSize } from 'src/hooks/suite';
+import { useAnalytics } from 'src/support/useAnalytics';
 
-import { EarnWithdrawalModalLoaded } from './EarnWithdrawalModalLoaded';
-import { useEarnModalAccount } from '../common/useEarnModalAccount';
+import { WithdrawalButton } from './WithdrawalForm/WithdrawalButton';
+import { WithdrawalForm } from './WithdrawalForm/WithdrawalForm';
+import { EarnWithdrawingInfo } from '../EarnInANutshell/components/EarnWithdrawingInfo';
 
 type EarnWithdrawalModalProps = {
     onCancel?: () => void;
-    account?: EarnAccountRef;
+    account: Account;
 };
 
-const createLoadedSelectedAccount = (account: Account): SelectedAccountLoaded => ({
-    status: 'loaded',
-    account,
-    network: getNetwork(account.symbol),
-    params: {
-        symbol: account.symbol,
-        accountIndex: account.index,
-        accountType: account.accountType,
-    },
-});
-
 export const EarnWithdrawalModal = ({ onCancel, account }: EarnWithdrawalModalProps) => {
-    const selectedAccount = useEarnModalAccount({
-        account,
-        shouldSyncSelectedAccount: true,
-    });
+    const analytics = useAnalytics();
+    const withdrawalContextValues = useWithdrawalForm({ account });
+    const { isBelowTablet } = useLayoutSize();
 
-    useEffect(() => {
-        if (!selectedAccount) {
-            onCancel?.();
-        }
-    }, [selectedAccount, onCancel]);
+    const onCancelClick = () => {
+        onCancel?.();
 
-    if (!selectedAccount) {
-        return null;
-    }
-
-    const loadedSelectedAccount = createLoadedSelectedAccount(selectedAccount);
+        analytics.report({
+            type: events.stakingUnstakeEvent.name,
+            payload: {
+                action: 'cancel',
+                step: 'unstake-form-modal',
+                networkSymbol: account.symbol,
+            },
+        });
+    };
 
     return (
-        <EarnWithdrawalModalLoaded onCancel={onCancel} selectedAccount={loadedSelectedAccount} />
+        <WithdrawalFormContext.Provider value={withdrawalContextValues}>
+            <Modal
+                width={960}
+                heading={
+                    <Translation
+                        id="TR_STAKE_UNSTAKE_TOKEN"
+                        values={{ symbol: getNetworkDisplaySymbol(account.symbol) }}
+                    />
+                }
+                description={
+                    account.networkType === 'cardano' ? (
+                        <Translation id="TR_STAKE_UNSTAKE_WITH_REWARDS" />
+                    ) : (
+                        <Translation id="TR_STAKE_CLAIM_AFTER_UNSTAKING" />
+                    )
+                }
+                onCancel={onCancelClick}
+                bottomContent={<WithdrawalButton />}
+            >
+                <Grid columns={isBelowTablet ? 1 : 2} gap={32}>
+                    <WithdrawalForm />
+                    <Column gap={20}>
+                        <CollapsibleBox
+                            heading={
+                                <H3 typographyStyle="body-md-strong">
+                                    <Translation id="TR_EARN_UNSTAKING_PROCESS" />
+                                </H3>
+                            }
+                            hasDivider={false}
+                            defaultIsOpen
+                        >
+                            <EarnWithdrawingInfo
+                                account={account}
+                                flow={EarnFlow.Stake}
+                                isExpanded
+                            />
+                        </CollapsibleBox>
+                    </Column>
+                </Grid>
+            </Modal>
+        </WithdrawalFormContext.Provider>
     );
 };
