+++ packages/suite/src/components/earn/utils/getEarnProviderName.ts
