+++ packages/suite/src/hooks/earn/useClaimForm.ts
@@ -2,8 +2,9 @@ import { createContext, useCallback, useContext, useEffect, useMemo } from 'reac
 import { useForm } from 'react-hook-form';
 
 import { getStakeFormsDefaultValues, getStakingContractAddress } from '@suite-common/staking';
+import { getNetwork } from '@suite-common/wallet-config';
 import { selectBaseCurrency, selectRawNetworkFeeInfo } from '@suite-common/wallet-core';
-import { PrecomposedTransactionFinal, SelectedAccountLoaded } from '@suite-common/wallet-types';
+import { Account, PrecomposedTransactionFinal } from '@suite-common/wallet-types';
 import { getConvertedOrDefaultFeeInfo } from '@suite-common/wallet-utils';
 
 import { signTransaction } from 'src/actions/wallet/stakeActions';
@@ -19,15 +20,15 @@ export const ClaimFormContext = createContext<ClaimContextValues | null>(null);
 ClaimFormContext.displayName = 'ClaimFormContext';
 
 type UseClaimFormsProps = {
-    selectedAccount: SelectedAccountLoaded;
+    account: Account;
 };
 
-export const useClaimForm = ({ selectedAccount }: UseClaimFormsProps): ClaimContextValues => {
+export const useClaimForm = ({ account }: UseClaimFormsProps): ClaimContextValues => {
     const dispatch = useDispatch();
 
     const baseCurrencyCode = useSelector(selectBaseCurrency);
 
-    const { account, network } = selectedAccount;
+    const network = getNetwork(account.symbol);
     const networkFees = useSelector(state => selectRawNetworkFeeInfo(state, account.symbol));
 
     const defaultValues = useMemo(() => {
