+++ packages/suite/src/hooks/earn/useSupplyForm.ts
@@ -5,17 +5,14 @@ import useDebounce from 'react-use/lib/useDebounce';
 import { fromWei } from 'web3-utils';
 
 import { getStakeFormsDefaultValues, getStakingContractAddress } from '@suite-common/staking';
+import { getNetwork } from '@suite-common/wallet-config';
 import {
     selectBaseCurrency,
     selectFiatRatesByFiatRateKey,
     selectRawNetworkFeeInfo,
     useFormDraft,
 } from '@suite-common/wallet-core';
-import {
-    PrecomposedTransactionFinal,
-    SelectedAccountLoaded,
-    StakeFormState,
-} from '@suite-common/wallet-types';
+import { Account, PrecomposedTransactionFinal, StakeFormState } from '@suite-common/wallet-types';
 import {
     fromBaseCurrencyToCryptoUnit,
     getConvertedOrDefaultFeeInfo,
@@ -40,13 +37,12 @@ export const SupplyFormContext = createContext<SupplyContextValues | null>(null)
 SupplyFormContext.displayName = 'SupplyFormContext';
 
 type UseSupplyFormProps = {
-    selectedAccount: SelectedAccountLoaded;
+    account: Account;
 };
 
-export const useSupplyForm = ({ selectedAccount }: UseSupplyFormProps): SupplyContextValues => {
+export const useSupplyForm = ({ account }: UseSupplyFormProps): SupplyContextValues => {
     const dispatch = useDispatch();
-
-    const { account, network } = selectedAccount;
+    const network = getNetwork(account.symbol);
 
     const baseCurrencyCode = useSelector(selectBaseCurrency);
     const networkFees = useSelector(state => selectRawNetworkFeeInfo(state, account.symbol));
