+++ packages/suite/src/views/wallet/staking/components/StakingDashboard/components/EverstakeFooter.tsx
@@ -3,13 +3,8 @@ import { useMemo } from 'react';
 import styled from 'styled-components';
 
 import { spacingsPx } from '@trezor/theme';
-import {
-    HELP_CENTER_ADA_STAKING,
-    HELP_CENTER_ETH_STAKING,
-    HELP_CENTER_SOL_STAKING,
-} from '@trezor/urls';
 
-import { PoweredByBadge } from 'src/components/earn';
+import { PoweredByBadge, getStakingHelpCenterLink } from 'src/components/earn';
 import { LearnMoreButton } from 'src/components/suite/LearnMoreButton';
 import { useSelector } from 'src/hooks/suite';
 import { selectSelectedAccount } from 'src/reducers/wallet/selectedAccountReducer';
@@ -27,18 +22,10 @@ const Wrapper = styled.div`
 export const EverstakeFooter = () => {
     const account = useSelector(selectSelectedAccount);
 
-    const learnMoreLink = useMemo(() => {
-        switch (account?.networkType) {
-            case 'ethereum':
-                return HELP_CENTER_ETH_STAKING;
-            case 'solana':
-                return HELP_CENTER_SOL_STAKING;
-            case 'cardano':
-                return HELP_CENTER_ADA_STAKING;
-            default:
-                return undefined;
-        }
-    }, [account]);
+    const learnMoreLink = useMemo(
+        () => getStakingHelpCenterLink(account?.networkType),
+        [account?.networkType],
+    );
 
     return (
         <Wrapper>
